Flatten three deep blocks and split the glyph outline walker out
FontRegistry.initialize held the whole load-one-font sequence inside two loops
inside a try; the sequence is register() now, and the loops just call it.

SpecShapes.paintItems was an if inside an if inside an if, building a list by
mutation. It is two named builders -- fillItem and strokeItem -- and a
listOfNotNull of them.

GlyphExtractor was eleven functions, five of which were the PathIterator walk
rather than anything about glyph extraction. Those, the Contour type and the four
constants only they use move to GlyphOutline.kt.

Styles dumped before and after: unchanged.
